# api/openapi.json
{
  "components": {
    "responses": {
      "BadRequest": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "The request was invalid."
      },
      "NotFound": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "The requested Spaniel resource was not found."
      }
    },
    "schemas": {
...
      "Error": {
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "Meta": {
        "additionalProperties": true,
        "properties": {
          "page": {
            "minimum": 1,
            "type": "integer"
          },
          "total": {
            "minimum": 0,
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "NotificationRecord": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "additionalProperties": false,
        "properties": {
          "acknowledged_at": {
            "type": [
              "number",
              "null"
            ]
          },
          "body": {
            "type": "string"
          },
          "created_at": {
            "type": "number"
          },
          "dedupe_key": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "link": {
            "type": "string"
          },
          "read_at": {
            "type": [
              "number",
              "null"
            ]
          },
          "severity": {
            "type": "string"
          },
          "source": {
            "type": "string"
          },
          "source_id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "source",
          "source_id",
          "severity",
          "title",
          "body",
          "link",
          "dedupe_key",
          "created_at",
          "read_at",
          "acknowledged_at"
        ],
        "type": "object"
      },
      "Ok": {
        "properties": {
          "ok": {
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "The complete local Spaniel HTTP API. JSON endpoints return the Envelope shape unless an operation explicitly returns an export.",
    "title": "Spaniel API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/notifications": {
      "get": {
        "operationId": "listNotifications",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "examples": {
                  "example": {
                    "summary": "Representative Spaniel response",
                    "value": {
                      "data": [
                        {
                          "body": "Representative Spaniel response",
                          "dedupe_key": "example",
                          "id": "session_2026-10-07_1200",
                          "link": "example",
                          "severity": "example",
                          "source": "example",
                          "source_id": "example",
                          "title": "example"
                        }
                      ],
                      "meta": {
                        "page": 1,
                        "total": 1
                      }
                    }
                  }
                },
                "schema": {
                  "properties": {
                    "data": {
                      "items": {
                        "$ref": "#/components/schemas/NotificationRecord"
                      },
                      "type": "array"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/Meta"
                    }
                  },
                  "required": [
                    "data",
                    "meta"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "summary": "list Notifications",
        "tags": [
          "Administration"
        ]
      }
    },
    "/api/notifications/{id}/acknowledge": {
      "post": {
        "operationId": "acknowledgeNotification",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "examples": {
                  "example": {
                    "summary": "Representative Spaniel response",
                    "value": {
                      "data": {
                        "ok": true
                      }
                    }
                  }
                },
                "schema": {
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Ok"
                    }
                  },
                  "required": [
                    "data"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "summary": "acknowledge Notification",
        "tags": [
          "Administration"
        ]
      }
    },
    "/api/notifications/{id}/read": {
      "post": {
        "operationId": "readNotification",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "examples": {
                  "example": {
                    "summary": "Representative Spaniel response",
                    "value": {
                      "data": {
                        "ok": true
                      }
                    }
                  }
                },
                "schema": {
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Ok"
                    }
                  },
                  "required": [
                    "data"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "summary": "read Notification",
        "tags": [
          "Administration"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Current Spaniel instance",
      "url": "/"
    }
  ],
...
}